+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* **2013-08-04**: Changed name of Sonata route names / patterns - now /admin/cmf/simplecms/foo instead of /admin/bundle/simplecms/foo
+
 1.0.0-beta3
 -----------